fixed symlinks
because git recognizes symlinks as symlinks and not their linked file,
had to reverse direction of link to get dotfile plan to work.

Now, original file is in this folder, and linked file is in place of
original.

NOTE: for push_dotfiles.sh to work, this folder must be located at:
$HOME/.dotfiles